Clarify testIdAttribute (#330)
I was a bit confused while reading through these docs as to whether or not the value I supplied to `testIdAttribute` would be prefixed with `data-`. It turns out that the value is _not_ prefixed, and so `my-data-test-id` wouldn't be a valid `data-` attribute. This change replaces the current example `my-data-test-id` with `data-my-test-id`.