Add which-query extension to documentation (#569)
* Add which-query extension to documentation

People come across the query page in the documentation, and they might not truly understand how to use `getByRole` or any other methods since they don't use to do that! They used to get an element by `class` name or `id` most of the time. Or they were simply using JQuery for many years. As a result, RTL's queries do not sound very familiar to them. The "which-query" chrome extension is a good place to help them get started. They can use it to learn how to access any element in the page, just by the help of RTL queries (instead of setting data-testid, ...)

* link to Testing Playground extension