docs: add findByText example
Add a simple JS login form to show examples of how to
use the `findBy*` query, in particular `findByText` for
DOM elements that may not be visible immediately.

Add two tests that assert for warnings that appear when the
user tries to submit the form with no username and password
and when the user enters invalid field values.