Fix broken links across several doc pages
As part of some recent documentation reorganisation, there are some
lingering links in the docs that point to the wrong page. There are
redirects set up that cover some of these, however, the redirects don't
appear to work for relative links or links with document anchors in
them. This commit updates:
- all references to `prettyDOM` now link to `api-debugging/#prettydom`
- all references to `getQueriesForElement` in `api-helpers` now link
  to `within` at `dom-testing-library/api-within`
- queries that previously linked to
  `dom/testing-library/api-queries` now link to `queries/about`
  (or a more relevant subpage for hash links)
- all references to `helpers` now link to `api-custom-queries`
- reference to `cleanup` in react-testing-library now links to
  `react-testing-library/api#cleanup`, as the only reference in
  `react-testing-library/setup` was how to skip it, which isn't what
  the text implied.