docs: update queries/about to explain `queryOptions`
* Update queries/about.mdx

`queryOptions` are mentioned, but it's never explained where you actually pass them in. Once you know it's obvious, but it wasn't initially to me ("do I pass it into the render function, to modify behavior for all queries?"), and readers shouldn't have to test until they find it.

It would be great to link to a list of the available options in `queryOptions` in the section I added, but I couldn't find such a list.

I also added a note aboout the parameters that functions passed to queries will receive. It can be derived from examples later in the document, but you have to look for them, and I think it's clearer to make it explicit when the option to pass functions is mentioned.

* Adjusting based on CR feedback

* Replacing code example for `queryOptions`

Replacing with link per CR suggestion.

* convert link to relative, descriptive link text

Co-authored-by: Matan Borenkraout <Matanbobi@gmail.com>

Co-authored-by: Matan Borenkraout <Matanbobi@gmail.com>